Refactor of focus mode: Moving code to new file cMain_Focus. Split funktionality of handlefocus mode between separate functions. Changed focus mode to look for the second to last save step instead of the last one.
