+++ src/GUI/GUI.fbp
@@ -8378,7 +8378,7 @@
           <property name="bitmap"></property>
           <property name="label">Steps</property>
           <property name="select">1</property>
-          <object class="wxPanel" expanded="false">
+          <object class="wxPanel" expanded="true">
             <property name="BottomDockable">1</property>
             <property name="LeftDockable">1</property>
             <property name="RightDockable">1</property>
@@ -8430,16 +8430,16 @@
             <property name="window_extra_style"></property>
             <property name="window_name"></property>
             <property name="window_style">wxTAB_TRAVERSAL</property>
-            <object class="wxBoxSizer" expanded="false">
+            <object class="wxBoxSizer" expanded="true">
               <property name="minimum_size"></property>
               <property name="name">step_panel_sizer</property>
               <property name="orient">wxVERTICAL</property>
               <property name="permission">none</property>
-              <object class="sizeritem" expanded="false">
+              <object class="sizeritem" expanded="true">
                 <property name="border">5</property>
                 <property name="flag">wxEXPAND</property>
                 <property name="proportion">1</property>
-                <object class="wxBoxSizer" expanded="false">
+                <object class="wxBoxSizer" expanded="true">
                   <property name="minimum_size"></property>
                   <property name="name">step_panel_search_sizer</property>
                   <property name="orient">wxHORIZONTAL</property>
@@ -8786,7 +8786,7 @@
                       <property name="style"></property>
                       <property name="subclass">; ; forward_declare</property>
                       <property name="toolbar_pane">0</property>
-                      <property name="tooltip">Focuses the list on steps on the last segments.&#x0A;By hidding every step before the last save point.</property>
+                      <property name="tooltip">Focuses the list on steps on the last segments.&#x0A;By hidding every step before the second to last save step.</property>
                       <property name="validator_data_type"></property>
                       <property name="validator_style">wxFILTER_NONE</property>
                       <property name="validator_type">wxDefaultValidator</property>
@@ -8799,7 +8799,7 @@
                   </object>
                 </object>
               </object>
-              <object class="sizeritem" expanded="false">
+              <object class="sizeritem" expanded="true">
                 <property name="border">5</property>
                 <property name="flag">wxEXPAND</property>
                 <property name="proportion">0</property>
@@ -9190,7 +9190,7 @@
                   </object>
                 </object>
               </object>
-              <object class="sizeritem" expanded="false">
+              <object class="sizeritem" expanded="true">
                 <property name="border">5</property>
                 <property name="flag">wxEXPAND</property>
                 <property name="proportion">1</property>
