+++ src/cMain.cpp
@@ -149,81 +149,6 @@ cMain::cMain() : GUI_Base(nullptr, wxID_ANY, window_title, wxPoint(30, 30), wxSi
 	TemplatePageStartup();
 }
 
-void cMain::OnStepsFocusCheckbox(wxCommandEvent& event)
-{
-	const int row_count = grid_steps->GetNumberRows();
-	auto selectedRows = grid_steps->GetSelectedRows();
-	const int first_row_index = grid_steps->IsSelection() ? selectedRows[0] : row_count - 1;
-
-	steps_focus_checkbox->Disable();
-	{
-		HandleFocusMode(event.IsChecked(), true);
-		if (first_row_index != -1)
-		{
-			grid_steps->GoToCell(row_count - 1, 0);
-			grid_steps->GoToCell(first_row_index - (first_row_index > 4 ? 3 : 0), 0); // move the grid to first selected row
-		}
-	}
-	steps_focus_checkbox->Enable();
-}
-
-void cMain::HandleFocusMode(bool checked, bool changed)
-{
-	const int row_count = grid_steps->GetNumberRows();
-
-	if (checked)
-	{
-		int last_save = 0;
-		for (int i = StepGridData.size() - 1; i >= 0; i--)
-		{
-			if (StepGridData[i].type == StepType::e_save && !StepGridData[i].Modifiers.skip)
-			{
-				last_save = i;
-				break;
-			}
-		}
-		if (last_save > 0)
-		{
-			grid_steps->BeginBatch();
-			{
-				for (int i = 0; i < last_save; i++)
-				{
-					grid_steps->HideRow(i);
-					if (changed && i % 800 == 799)
-					{
-						grid_steps->EndBatch();
-						this->Update();
-						grid_steps->BeginBatch();
-					}
-				}
-			}
-			grid_steps->EndBatch();
-		
-
-			for (int i = last_save; i < row_count; i++)
-				grid_steps->ShowRow(i);
-		}
-	}
-	else
-	{
-		grid_steps->BeginBatch();
-		{
-			for (int i = 0; i < row_count; i++)
-			{
-				grid_steps->ShowRow(i);
-				if (changed && i % 800 == 799)
-				{
-					grid_steps->GoToCell(row_count - 1, 0);
-					grid_steps->EndBatch();
-					this->Update();
-					grid_steps->BeginBatch();
-				}
-			}
-		}
-		grid_steps->EndBatch();
-	}
-}
-
 void cMain::OnStepsGridCellChange(wxGridEvent& event)
 {
 	auto row = event.GetRow();
